Add validation and `HasValue` support to delete handlers
Refactored `DeleteElementCommandHandler` and `ElementTypeDeleteCommandHandler`
to validate entity existence using `IMediator` and `HasValue`. Updated both
handlers to be asynchronous.

Enhanced `IUnitResult` with a `HasValue` property to indicate valid results.
Modified `UnitResult` to implement `HasValue` based on `IsSuccess` and
non-null `Result`.

These changes improve robustness by ensuring entities exist before deletion
and provide clearer result validation in the `IUnitResult` abstraction.

+++ CF.GameEngine/Web/CF.GameEngine.Web.Api/Features/Element/Delete/DeleteElementCommandHandler.cs
@@ -1,13 +1,23 @@
 ﻿using CF.GameEngine.Infrastructure.Features.Elements;
+using CF.GameEngine.Web.Api.Features.Element.Get;
 using IDFCR.Shared.Abstractions.Results;
+using IDFCR.Shared.Exceptions;
 using IDFCR.Shared.Mediatr;
+using MediatR;
 
 namespace CF.GameEngine.Web.Api.Features.Element.Delete;
 
-public class DeleteElementCommandHandler(IElementRepository elementRepository) : IUnitRequestHandler<DeleteElementCommand>
+public class DeleteElementCommandHandler(IMediator mediator, IElementRepository elementRepository) : IUnitRequestHandler<DeleteElementCommand>
 {
-    public Task<IUnitResult> Handle(DeleteElementCommand request, CancellationToken cancellationToken)
+    public async Task<IUnitResult> Handle(DeleteElementCommand request, CancellationToken cancellationToken)
     {
-        return elementRepository.DeleteAsync([request.ElementId], cancellationToken);
+        var element = await mediator.Send(new ElementFindByIdQuery(request.ElementId), cancellationToken);
+
+        if (!element.HasValue)
+        {
+            return new UnitResult(new EntityNotFoundException(typeof(ElementDto), request.ElementId));
+        }
+
+        return await elementRepository.DeleteAsync([request.ElementId], cancellationToken);
     }
 }
